Asteriods
Calculating when the maximum overlap occurs for two asteroids requires a
bit of programming.
The input consists of two asteroid specications. Each has the form
n x1 y1 x2 y2 : : : xn yn vx vy
where n(3  n  10) is the number of vertices, each xi; yi(􀀀10000  xi;
yi  10000) are the
coordinates of a vertex of the asteroid on Hans screen given in
clockwise order, and vx; vy(􀀀100 
vx; vy  100) are the x and y velocities (in units/second) of the
asteroid. The xi; yi values specify
the location of each asteroid at time t = 0, and the polygons do not
intersect or touch at this time.
The maximum length of any side of an asteroid is 500. All numbers in the
input are integers.